backend/requirements.txt
appnope==0.1.0
asgiref==3.2.3
autopep8==1.5
backcall==0.1.0
blessings==1.7
boto3==1.11.1
botocore==1.14.17
Click==7.0
coverage==5.0.3
decorator==4.4.1
dj-database-url==0.5.0
Django==3.0.2
django-configurations==2.2
django-cors-headers==3.2.1
django-filter==2.2.0
django-model-utils==4.0.0
django-nose==1.4.6
django-storages==1.8
django-unique-upload==0.2.1
djangorestframework==3.11.0
docutils==0.15.2
entrypoints==0.3
factory-boy==2.12.0
Faker==4.0.0
flake8==3.7.9
gunicorn==20.0.4
ipdb==0.12.3
ipython==7.11.1
ipython-genutils==0.2.0
jedi==0.16.0
Jinja2==2.11.1
jmespath==0.9.4
livereload==2.6.1
Markdown==3.1.1
MarkupSafe==1.1.1
mccabe==0.6.1
mkdocs==1.0.4
mock==3.0.5
newrelic==5.4.1.134
nose==1.3.7
nose-progressive==1.5.2
parso==0.6.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.3
psycopg2-binary==2.8.4
ptyprocess==0.6.0
pycodestyle==2.5.0
pyflakes==2.1.1
Pygments==2.5.2
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3
s3transfer==0.3.3
six==1.14.0
sqlparse==0.3.0
text-unidecode==1.3
tornado==6.0.3
traitlets==4.3.3
urllib3==1.25.8
wcwidth==0.1.8
yapf==0.29.0
